% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\docType{data}
\name{binary_exts}
\alias{binary_exts}
\title{Vector of common binary file extensions (svg isn't really binary but for our purposes...)}
\format{An object of class \code{character} of length 18.}
\usage{
binary_exts
}
\description{
Vector of common binary file extensions (svg isn't really binary but for our purposes...)
}
\keyword{datasets}
