% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_counts.R
\name{count_lines_ext}
\alias{count_lines_ext}
\title{count the lines in all the files with a given extension}
\usage{
count_lines_ext(filenames, ext)
}
\arguments{
\item{filenames}{vector of files to search for files with the extension}

\item{ext}{a file extension eg "txt".  Should be for text files of some sort}
}
\description{
count the lines in all the files with a given extension
}
\examples{
lf <- list.files(recursive = TRUE, full.names = TRUE)
count_lines_ext(lf, "r")
count_lines_ext(lf, "sql")
}
