% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_counts.R
\name{count_lines_ext_v}
\alias{count_lines_ext_v}
\title{count the lines in all the files with a vector of given extensions}
\usage{
count_lines_ext_v(filenames, exts, label = NULL, onlyknown = TRUE)
}
\arguments{
\item{filenames}{a vector of full path file names to search for files of those extensions}

\item{exts}{a vector of file extensions eg c("txt", "r", "sas", "c")}

\item{label}{a potential extra column to add, handy if using this function on many different folder systems}

\item{onlyknown}{should the program only proceed if it recognises all the extensions as text files?}
}
\description{
count the lines in all the files with a vector of given extensions
}
\examples{
lf <- list.files(recursive = TRUE, full.names = TRUE)
count_lines_ext_v(lf, text_exts)
count_files_ext_v(lf, binary_exts)
}
