% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_exts.R
\docType{data}
\name{data_exts}
\alias{data_exts}
\title{Vector of common data file extensions}
\format{An object of class \code{character} of length 5.}
\usage{
data_exts
}
\description{
Vector of common data file extensions
}
\keyword{datasets}
