% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_exts.R
\docType{data}
\name{exe_exts}
\alias{exe_exts}
\title{Vector of file extensions for actual computer programs}
\format{An object of class \code{character} of length 3.}
\usage{
exe_exts
}
\description{
Vector of file extensions for actual computer programs
}
\keyword{datasets}
