% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_counts.R
\name{folder_file_counts}
\alias{folder_file_counts}
\title{File counts in multiple folders}
\usage{
folder_file_counts(foldernames, exts)
}
\arguments{
\item{foldernames}{a vector of paths to folders to search (with their subdirectories) for files}

\item{exts}{a vector of file extensions eg c("r", "sas", "sql") to find and count in each folder system}
}
\description{
Count all the files of given extension, in each of a collection of folder systems
}
