% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder_counts.R
\name{folder_line_counts}
\alias{folder_line_counts}
\title{Line counts in multiple folders}
\usage{
folder_line_counts(foldernames, exts = text_exts)
}
\arguments{
\item{foldernames}{a vector of paths to folders to search (with their subdirectories) for files}

\item{exts}{a vector of file extensions eg c("r", "sas", "sql") to find and count in each folder system}
}
\description{
Count the lines in all the files of given extensions, in each of a collection of folder systems
}
