% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclouds.R
\name{folder_wordcloud}
\alias{folder_wordcloud}
\title{Count words in a target folder system}
\usage{
folder_wordcloud(foldername, plot = TRUE, ...)
}
\arguments{
\item{foldername}{path to a folder or directory that contains some text files}

\item{plot}{should the word cloud be drawn}

\item{...}{other parameters to pass to wordcloud}
}
\description{
Counts words and draws a word cloud of all files in a folder system
}
