% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_exts.R
\docType{data}
\name{image_exts}
\alias{image_exts}
\alias{image_exts}
\title{Vector of common image file extensions}
\format{An object of class \code{character} of length 3.}
\usage{
image_exts

image_exts
}
\description{
Vector of common image file extensions

Vector of executable file extensions
}
\keyword{datasets}
