% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordclouds.R
\name{MakeSpaces}
\alias{MakeSpaces}
\title{utility function, not for export, converting some common character combinations to spaces}
\usage{
MakeSpaces(x, From = c("(", ")", "$", "{", "}", ",", "<-", "=", "+", "*", "[",
  "]", ":", "\\"", "'", "#", "\%", ":", ".", ";"))
}
\arguments{
\item{x}{character vector}

\item{From}{characters to turn into spaces}
}
\description{
utility function, not for export, converting some common character combinations to spaces
}
