% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_exts.R
\docType{data}
\name{text_exts}
\alias{text_exts}
\title{Vector of common text file extensions}
\format{An object of class \code{character} of length 15.}
\usage{
text_exts
}
\description{
Vector of common text file extensions
}
\keyword{datasets}
