% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\docType{data}
\name{textfile_exts}
\alias{textfile_exts}
\title{Vector of common text file extensions}
\format{An object of class \code{character} of length 9.}
\usage{
textfile_exts
}
\description{
Vector of common text file extensions
}
\keyword{datasets}
