% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_counts.R
\name{count_files_ext_v}
\alias{count_files_ext_v}
\title{Count files}
\usage{
count_files_ext_v(filenames, exts, label = NULL)
}
\arguments{
\item{filenames}{vector of full path filenames to search}

\item{exts}{vector of file extensions eg c("doc", "docx", "ppt")}

\item{label}{a potential extra element to add, handy if using this function on many different folder systems}
}
\value{
a data frame with a single row and a column for each file extension, with counts as values.  
If label is not NULL, there will be an additional column label.
}
\description{
count the files, given a vector of extensions
}
\examples{
lf <- list.files(recursive = TRUE, full.names = TRUE)
count_files_ext_v(lf, binary_exts)
count_files_ext_v(lf, text_exts)
}
