% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counting.R
\name{count_files_ext_v}
\alias{count_files_ext_v}
\title{Count files}
\usage{
count_files_ext_v(exts, filenames)
}
\arguments{
\item{exts}{vector of file extensions eg c("doc", "docx", "ppt")}

\item{filenames}{vector of full path filenames to search}
}
\description{
count the files, given a vector of extensions
}
\examples{
lf <- list.files(recursive = TRUE, full.names = TRUE)
count_files_ext_v(binary_exts, lf)
count_files_ext_v(textfile_exts, lf)
}
