% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_exts.R
\docType{data}
\name{image_exts}
\alias{image_exts}
\title{Vector of common image file extensions}
\format{An object of class \code{character} of length 6.}
\usage{
image_exts
}
\description{
Vector of common image file extensions
}
\keyword{datasets}
